\name{internal}

\title{Internal functions used in package `predmixcor'}

\description{this function is internal. Type the function name to see its definition}

\alias{LOG.add.exp}

\alias{LOG.prob} 

\alias{LOG.sqdiff.exp} 

\alias{LOG.sum.exp} 

\alias{Predict.mix} 

\alias{abs_cor}

\alias{bound} 

\alias{countN} 

\alias{gen.alpha.IGamma} 

\alias{indexconvert} 

\alias{lf.phi} 

\alias{lf.r} 

\alias{mh1.unif}

\alias{present} 

\alias{prob.add} 

\alias{prob.cor} 

\alias{prob.cor.sum} 

\alias{prob.sch} 

\alias{selfth.abscor} 

\alias{simpson}  
  
\alias{training}

\keyword{internal}
